+++ Secp256k1.Net/Secp256k1.cs
@@ -376,7 +376,7 @@ namespace Secp256k1Net
                 throw new ArgumentException($"{nameof(signatureOutput)} must be {SERIALIZED_DER_SIGNATURE_MAX_SIZE} bytes as maximum to void truncate signature");
             }
 
-            uint sigOutputLength = (uint)SERIALIZED_DER_SIGNATURE_MAX_SIZE;
+            ulong sigOutputLength = (ulong)SERIALIZED_DER_SIGNATURE_MAX_SIZE;
             
             fixed (byte* sig = &MemoryMarshal.GetReference(signatureOutput),
                 input = &MemoryMarshal.GetReference(signatureInput))
